server/requirements.txt
Flask>=0.10
psycopg2
beautifulsoup4
requests
SQLAlchemy>=0.9
GeoAlchemy>=0.7
Flask-restful>=0.2
Flask-sqlalchemy>=2.0
python-Levenshtein
